+++ packages/gatsby-theme-api-docs/gatsby-config.js
@@ -1,6 +1,18 @@
 const path = require('path');
 
+const getTagListOption = (themeOptions) => {
+  if (!themeOptions.transformerMdx) return [];
+  return (
+    themeOptions.transformerMdx.tagList ||
+    // backwards compatibility
+    themeOptions.transformerMdx.tagWhitelist
+  );
+};
+
 module.exports = (themeOptions = {}) => {
+  // Extract the `tagList` from the theme options if specified
+  const additionalTags = getTagListOption(themeOptions);
+
   return {
     plugins: [
       {
@@ -8,6 +20,13 @@ module.exports = (themeOptions = {}) => {
         options: {
           name: 'api-specs',
           path: path.resolve('./src/api-specs'),
+        },
+      },
+      {
+        resolve: '@commercetools-docs/gatsby-transformer-mdx-introspection',
+        options: {
+          ...themeOptions.transformerMdx,
+          tagList: ['ApiType', 'ApiEndpoint', ...additionalTags],
         },
       },
       {