+++ packages/gatsby-theme-docs/src/components/card.js
@@ -1,6 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { cardElements } from '@commercetools-docs/ui-kit';
+import styled from '@emotion/styled';
+import {
+  Markdown,
+  markdownFragmentToReact,
+  cardElements,
+} from '@commercetools-docs/ui-kit';
 import GatsbyLink from './link';
 
 const {
@@ -15,6 +20,24 @@ const {
   Title,
 } = cardElements;
 
+const BodyContent = (props) => {
+  if (typeof props.children === 'string') {
+    return props.clickable
+      ? markdownFragmentToReact(props.children, { a: styled.span`` })
+      : markdownFragmentToReact(props.children, { a: GatsbyLink });
+  }
+  return (
+    <Markdown.TypographyContainer>
+      {props.children}
+    </Markdown.TypographyContainer>
+  );
+};
+BodyContent.displayName = 'BodyContent';
+BodyContent.propTypes = {
+  clickable: PropTypes.bool,
+  children: PropTypes.oneOfType([PropTypes.string, PropTypes.node]).isRequired,
+};
+
 const WrapWith = (props) =>
   props.condition ? props.wrapper(props.children) : props.children;
 WrapWith.displayName = 'WrapWith';
@@ -22,68 +45,6 @@ WrapWith.propTypes = {
   condition: PropTypes.bool.isRequired,
   wrapper: PropTypes.func.isRequired,
   children: PropTypes.node.isRequired,
-};
-
-const isReactLink = (alienElement) =>
-  React.isValidElement(alienElement) && alienElement.props.href;
-
-/**
- * Expects a React Link element, returns just the text contained in the link.
- * It handles only the simple case where the link text is a simple string.
- */
-const removeLink = (linkNode) => {
-  if (isReactLink(linkNode)) {
-    if (typeof linkNode.props.children === 'string') {
-      return linkNode.props.children;
-    }
-    return '';
-  }
-};
-
-/**
- * This component main purpose is to manage the case in which a clickable card has
- * a body with a link defined. This is an invalid case as you'll end up with 2 nesed links.
- * So the processChildren function run through the body content removing any link and just
- * leaving a plain string (non linked).
- */
-const BodyContent = (props) => {
-  if (props.clickable) {
-    return processChildren(props.children);
-  }
-  return props.children;
-};
-
-BodyContent.displayName = 'BodyContent';
-BodyContent.propTypes = {
-  clickable: PropTypes.bool,
-  children: PropTypes.node.isRequired,
-};
-
-const processChildren = (childrenNode) => {
-  // it's just a string... phew!
-  if (typeof childrenNode === 'string') {
-    return childrenNode;
-  }
-  // React component witch is a Link
-  if (isReactLink(childrenNode)) {
-    return removeLink(childrenNode);
-  }
-  // any other React component
-  if (React.isValidElement(childrenNode)) {
-    return React.cloneElement(
-      childrenNode,
-      undefined,
-      processChildren(childrenNode.props.children)
-    );
-  }
-  // array
-  if (Array.isArray(childrenNode)) {
-    const result = [];
-    childrenNode.forEach((nodeItem) => {
-      result.push(processChildren(nodeItem));
-    });
-    return result;
-  }
 };
 
 const Card = (props) => (