+++ packages/gatsby-theme-docs/src/templates/homepage.js
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { IntlProvider } from 'react-intl';
 import { graphql } from 'gatsby';
+import { MDXRenderer } from 'gatsby-plugin-mdx';
 import { MDXProvider } from '@mdx-js/react';
 import { Markdown } from '@commercetools-docs/ui-kit';
 import { PageDataContext } from '../hooks/use-page-data';
@@ -23,7 +24,9 @@ const HomepageTemplate = (props) => (
           <MDXProvider components={markdownComponents}>
             <Markdown.TypographyPage>
               {/* This wrapper div is important to ensure the vertical space */}
-              <div>{props.children}</div>
+              <div>
+                <MDXRenderer>{props.data.contentPage.body}</MDXRenderer>
+              </div>
             </Markdown.TypographyPage>
           </MDXProvider>
         </LayoutContentHomepage>
@@ -46,12 +49,12 @@ HomepageTemplate.propTypes = {
       beta: PropTypes.bool.isRequired,
       excludeFromSearchIndex: PropTypes.bool.isRequired,
       allowWideContentLayout: PropTypes.bool.isRequired,
+      body: PropTypes.string.isRequired,
     }).isRequired,
     heroBackground: PropTypes.shape({
       publicURL: PropTypes.string.isRequired,
     }).isRequired,
   }).isRequired,
-  children: PropTypes.any.isRequired,
 };
 export default HomepageTemplate;
 
@@ -78,6 +81,7 @@ export const query = graphql`
       beta
       excludeFromSearchIndex
       allowWideContentLayout
+      body
     }
     heroBackground: file(relativePath: { eq: $heroBackgroundRelativePath }) {
       publicURL