+++ packages/gatsby-theme-docs/src/templates/release-notes-detail.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { IntlProvider } from 'react-intl';
 import { graphql } from 'gatsby';
 import { MDXProvider } from '@mdx-js/react';
+import { MDXRenderer } from 'gatsby-plugin-mdx';
 import { Markdown } from '@commercetools-docs/ui-kit';
 import LayoutReleaseNotesDetail from '../layouts/release-notes-detail';
 import LayoutReleaseNoteBody from '../layouts/internals/layout-release-note-body';
@@ -28,7 +29,7 @@ const ReleaseNotesDetailTemplate = (props) => (
           <Markdown.TypographyPage>
             <div>
               <LayoutReleaseNoteBody {...props.data.releaseNotePage}>
-                {props.children}
+                <MDXRenderer>{props.data.releaseNotePage.body}</MDXRenderer>
               </LayoutReleaseNoteBody>
             </div>
           </Markdown.TypographyPage>
@@ -50,7 +51,6 @@ ReleaseNotesDetailTemplate.propTypes = {
       body: PropTypes.string.isRequired,
     }).isRequired,
   }).isRequired,
-  children: PropTypes.any.isRequired,
 };
 
 export default ReleaseNotesDetailTemplate;