+++ packages/ui-kit/src/mermaid.d.ts
@@ -1,4 +1,0 @@
-// unfortunately the DefinitelyTyped @types/mermaid interfaces are
-// so incomplete in the theming config that it's not possible to use them
-// when intensively theming like here. So the mermaid API is used untyped
-declare module 'mermaid';