+++ packages/gatsby-theme-docs/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @commercetools-docs/gatsby-theme-docs
 
+## 21.0.0
+
+### Major Changes
+
+- [#1535](https://github.com/commercetools/commercetools-docs-kit/pull/1535) [`c8d5b734`](https://github.com/commercetools/commercetools-docs-kit/commit/c8d5b734603ba77bfd36c84b7f4ac3e53dab35d2) Thanks [@gabriele-ct](https://github.com/gabriele-ct)! - Revert all the changes strictly related with MDX v2 because of performance issue on local development. This version includes all the refactoring and performance optimizations implemented during MDX v2 implementation.
+
+### Minor Changes
+
+- [#1530](https://github.com/commercetools/commercetools-docs-kit/pull/1530) [`0624304b`](https://github.com/commercetools/commercetools-docs-kit/commit/0624304b519b50c406d254d1a1b2636124a32bbf) Thanks [@nkuehn](https://github.com/nkuehn)! - LOW_MEM=true flag also activates experimental webpack lazy compilation in development server.
+  This makes the development server throw errors that require reloading it but the effect on startup time and memory is extremely high.
+  Purpose is to allow testing MDX v2 plugin based work.
+
+### Patch Changes
+
+- Updated dependencies [[`c8d5b734`](https://github.com/commercetools/commercetools-docs-kit/commit/c8d5b734603ba77bfd36c84b7f4ac3e53dab35d2), [`28f404f3`](https://github.com/commercetools/commercetools-docs-kit/commit/28f404f377b79939bf58c039ef166c1000198a92)]:
+  - @commercetools-docs/ui-kit@21.0.0
+
 ## 20.2.0
 
 ### Minor Changes